% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrays-core.R
\name{aInsert}
\alias{aInsert}
\title{Array Slice Insertion}
\usage{
aInsert(x, d, values, replacement, index = F)
}
\arguments{
\item{x}{An array, a vector}

\item{d}{Numeric or character. The dimension to subset.}

\item{values}{Numeric or character vector. The values to subset.}

\item{replacement}{Replacement for selected subset}

\item{index}{Should numeric values be treated as indexes ?}
}
\value{
An array or vector with modified subset over one dimension.
}
\description{
Slice an array and modify values from a subset over one of the dimensions
}
