% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrays-core.R
\name{aMake}
\alias{aMake}
\title{Quick Array Creation}
\usage{
aMake(data, dimnames)
}
\arguments{
\item{data}{The array data}

\item{dimnames}{The array dimension names}
}
\value{
An array
}
\description{
Create an Array without providing the dim argument
}
