% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrays-core.R
\name{aMerge}
\alias{aMerge}
\title{Array Mergeing}
\usage{
aMerge(l, varname = "Source", fdims = NULL)
}
\arguments{
\item{l}{A list of arrays}

\item{varname}{The name to be given to the variable that contains the array
source}

\item{fdims}{The dimensions for the final array, minus the array source. In
case this argument is not provided, summation is performed over dimensions
not present in at least one array, then all indices present in at least one
array are included.}
}
\value{
An agregated array
}
\description{
Merge a list of arrays
}
