% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrays-core.R
\name{aMorph}
\alias{aMorph}
\title{Array Transformation}
\usage{
aMorph(x, l, keep = TRUE, reorder = !keep)
}
\arguments{
\item{x}{A vector or an array}

\item{l}{A list whose components are named after some of the array dimensions
and contain the indices to keep}

\item{keep}{Boolean. Should the dimensions that are not in the list be summed
over ?}

\item{reorder}{Boolean. Should the array dimensions be permuted so that they
match the order of the list ?}
}
\value{
An array containing only the specified components
}
\description{
Transform an array, keeping the desired indices and dimensions
}
