% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrays-core.R
\name{array_to_df}
\alias{array_to_df}
\title{Array to data.frame conversion}
\usage{
array_to_df(data, value.name)
}
\arguments{
\item{data}{An array}

\item{value.name}{The name of the column to store the array values.}
}
\value{
A tibble
}
\description{
Convert an array object to a tibble
}
