% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrays-package.R
\docType{package}
\name{arrays-package}
\alias{arrays-package}
\title{arrays: Convenience functions for the manipulation of arrays}
\description{
The arrays package contains 6 functions for the manipulation of arrays and
one function that converts a data.frame into an array
}
