% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrays-core.R
\name{aSlice}
\alias{aSlice}
\title{Array Slicing}
\usage{
aSlice(x, d, values, index = F)
}
\arguments{
\item{x}{An array, a vector}

\item{d}{Numeric or character. The dimension to subset.}

\item{values}{Numeric or character vector. The values to subset.}

\item{index}{Should numeric values be treated as indexes ?}
}
\value{
An array or vector subset over one dimension.
}
\description{
Slice an array, returning a subset over one of the dimensions
}
