% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrays-core.R
\name{aSums}
\alias{aSums}
\title{Array Sum over Dimensions}
\usage{
aSums(x, d, keep = TRUE, na.rm = FALSE)
}
\arguments{
\item{x}{An array}

\item{d}{Numeric or character vector. The dimensions to be considered.}

\item{keep}{Boolean. Should those dimensions be kept or dropped in the final
result. By default the sum is made over all dimensions that are not
explicitely named.}

\item{na.rm}{Should NA value be removed ? Default : FALSE}
}
\value{
An array or vector containing the sum over the undesired dimensions
}
\description{
Return an array summed over one or several dimensions
}
