% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrays-core.R
\name{df_to_array}
\alias{df_to_array}
\title{Data.frame to array conversion}
\usage{
df_to_array(data, covariates, value.var, fill = 0)
}
\arguments{
\item{data}{A data.frame.}

\item{covariates}{Character vector. The covariates to use for array
dimensions. Unselected dimensions will be collapsed.}

\item{value.var}{Character. The name of the column which contains the value
for the array.}

\item{fill}{Numeric. The value to use when the combination of covariates does
not exist in the data.frame}
}
\value{
An array whose dimensions correspond to the selected covariates.
}
\description{
Convert a data.frame object to an array
}
