% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrays-package.R
\docType{package}
\name{arrays-package}
\alias{arrays-package}
\title{arrays: Convenience functions for the manipulation of arrays}
\description{
The arrays package contains 7 functions for the manipulation of arrays and
conversion between data.frame (tibble) and arrays.
}
